pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-memory"
version = "1.0.0"
description = "OpenClaw 本地语义记忆与检索系统 — ONNX + FAISS"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "QClaw / OpenClaw Community"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
    "Operating System :: OS Independent",
]
keywords = ["semantic-memory", "embedding", "faiss", "onnx", "openclaw", "vector-search", "rag"]

dependencies = [
    "onnxruntime>=1.17.0",
    "tokenizers>=0.19.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0,<2",
    "chardet>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "apscheduler>=3.10.0",
]
docs = [
    "python-docx>=1.0",
    "PyPDF2>=3.0",
    "cryptography>=42.0",
    "openpyxl>=3.1",
    "python-pptx>=1.0",
]
all = [
    "semantic-memory[dev,server,docs]",
]

[project.scripts]
semantic-memory = "run:main"
semantic-memory-server = "api_server:run"

[project.urls]
Homepage = "https://github.com/zhan1206/semantic-memory"
Documentation = "https://github.com/zhan1206/semantic-memory#readme"
Repository = "https://github.com/zhan1206/semantic-memory"
Issues = "https://github.com/zhan1206/semantic-memory/issues"

[tool.setuptools.packages.find]
where = ["scripts"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["scripts"]
omit = ["tests/*", "api_server.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]